requirements.txt
django>=5.2.11

# django
whitenoise
gunicorn            # The production web server
psycopg2-binary     # PostgreSQL adapter
dj-database-url     # To parse the database URL provided by Seenode
# whitenoise          # To serve Admin Panel CSS/Static files
python-dotenv
django-jazzmin==2.6.2
django-modeladmin-reorder